epam6/requirements.txt
flake8==3.9.2
pycodestyle==2.7.0
pytest==7.1.2
tox==3.23.1
jinja2
PyYAML